Combine dragger into Nav
moved recording status to left and put window controls to right
